+++ lib/src/cache/sync_cache.dart
@@ -1,5 +1,7 @@
 import 'dart:async';
 
+import 'package:ecache/src/storage/storage_mgr.dart';
+
 import '../../ecache.dart';
 
 /// A generic, stripped down class for [Cache] implementations if you do not need async calls.
@@ -24,9 +26,14 @@ class SyncCache<K, V> extends Cache<K, V> {
   /// An optional [strategy] can be provided. If not, a [SimpleStrategy]
   /// instance is used.
   SyncCache({Storage<K, V>? storage, required int capacity, AbstractStrategy<K, V>? strategy})
-      : storage = storage ?? SimpleStorage<K, V>(),
+      : storage = storage ?? (StorageMgr().isEnabled() ? StatisticsStorage<K, V>() : SimpleStorage<K, V>()),
         strategy = strategy ?? SimpleStrategy<K, V>() {
     this.strategy.init(this.storage, capacity);
+  }
+
+  @override
+  void dispose() {
+    storage.dispose();
   }
 
   /// Synchronously returns the element for the given [key], or `null` if the key is not found.