+++ lib/src/storage/statistics_storage.dart
@@ -1,32 +1,28 @@
-import 'dart:math';
+import 'package:ecache/src/storage/storage_mgr.dart';
 
 import '../../ecache.dart';
 
 /// Same as [SimpleStorage] but collects a few statistical data.
 class StatisticsStorage<K, V> extends SimpleStorage<K, V> {
-  int _maxLength = 0;
-  int _hitCount = 0;
-  int _missCount = 0;
-  int _evictionCount = 0;
+  final StorageMetric storageMetric = StorageMetric();
 
-  /// The number of times a requested item was found in the cache.
-  int get hitCount => _hitCount;
+  StatisticsStorage({super.onEvict}) {
+    StorageMgr().register(this);
+  }
 
-  /// The number of times a requested item was not found in the cache.
-  int get missCount => _missCount;
-
-  /// The number of times an item was evicted from the cache to make space.
-  int get evictionCount => _evictionCount;
-
-  StatisticsStorage({super.onEvict});
+  @override
+  void dispose() {
+    StorageMgr().unregister(this);
+    clear();
+  }
 
   @override
   CacheEntry<K, V>? get(K key) {
     final entry = super.get(key);
     if (entry != null) {
-      _hitCount++;
+      storageMetric.incHitCount();
     } else {
-      _missCount++;
+      storageMetric.incMissCount();
     }
     return entry;
   }
@@ -34,20 +30,21 @@ class StatisticsStorage<K, V> extends SimpleStorage<K, V> {
   @override
   void setInternal(K key, CacheEntry<K, V> value) {
     super.setInternal(key, value);
-    _maxLength = max(_maxLength, length);
+    storageMetric.checkLength(length);
+    storageMetric.incSetCount();
   }
 
   @override
   CacheEntry<K, V>? onCapacity(K key) {
     final entry = super.onCapacity(key);
     if (entry != null) {
-      _evictionCount++;
+      storageMetric.incEvictionCount();
     }
     return entry;
   }
 
   @override
   String toString() {
-    return 'StatisticsStorage{current: $length, max: $_maxLength, hits: $_hitCount, misses: $_missCount, evictions: $_evictionCount}';
+    return 'StatisticsStorage{storageMetric: $storageMetric}';
   }
 }